+++ src/components/Molecules/Header/Header.md
@@ -2,5 +2,17 @@
 
 ```js
 import data from './data/data';
-<Header items={data} color="green" />;
+import Link from '../../Atoms/Link/Link';
+
+<Header navItems={data}>
+  <Link color="green" href="/donation">
+    Donate
+  </Link>
+  <Link color="red" href="/donation">
+    Donate
+  </Link>
+  <Link color="black" href="/donation">
+    Donate
+  </Link>
+</Header>;
 ```
